+++ sample-apps/react-native/dogfood/src/components/MeetingUI.tsx
@@ -13,6 +13,7 @@ import { useAppGlobalStoreSetState } from '../contexts/AppContext';
 import { AuthenticationProgress } from './AuthenticatingProgress';
 import { CallErrorComponent } from './CallErrorComponent';
 import { useUnreadCount } from '../hooks/useUnreadCount';
+import { LayoutProvider } from '../contexts/LayoutContext';
 
 type Props = NativeStackScreenProps<
   MeetingStackParamList,
@@ -60,6 +61,7 @@ export const MeetingUI = ({ callId, navigation, route }: Props) => {
     try {
       call?.updatePublishOptions({
         preferredCodec: 'vp9',
+        preferredBitrate: 1500000,
       });
       await call?.join({ create: true });
       appStoreSetState({ chatLabelNoted: false });
@@ -128,13 +130,14 @@ export const MeetingUI = ({ callId, navigation, route }: Props) => {
     );
   } else {
     return (
-      <ActiveCall
-        onCallEnded={onCallEnded}
-        onChatOpenHandler={onChatOpenHandler}
-        unreadCountIndicator={unreadCountIndicator}
-        onHangupCallHandler={onHangupCallHandler}
-        onBackPressed={onHangupCallHandler}
-      />
+      <LayoutProvider>
+        <ActiveCall
+          onCallEnded={onCallEnded}
+          onHangupCallHandler={onHangupCallHandler}
+          onChatOpenHandler={onChatOpenHandler}
+          unreadCountIndicator={unreadCountIndicator}
+        />
+      </LayoutProvider>
     );
   }
 };