+++ sample-apps/react-native/dogfood/src/contexts/AppContext.tsx
@@ -1,7 +1,8 @@
+import { REACT_NATIVE_DOGFOOD_APP_ENVIRONMENT } from '@env';
 import createStoreContext from './createStoreContext';
 
 export type AppMode = 'Meeting' | 'Call' | 'Audio-Room' | 'LiveStream' | 'None';
-export type AppEnvironment = 'pronto' | 'demo';
+export type ThemeMode = 'dark' | 'light';
 
 type AppGlobalStore = {
   apiKey: string;
@@ -11,6 +12,9 @@ type AppGlobalStore = {
   appMode: AppMode;
   appEnvironment: AppEnvironment;
   chatLabelNoted?: boolean;
+  themeMode: ThemeMode;
+  localIpAddress: string;
+  useLocalSfu?: boolean;
 };
 
 export const {
@@ -24,8 +28,20 @@ export const {
     userImageUrl: '',
     userName: '',
     appMode: 'None',
-    appEnvironment: 'demo',
+    appEnvironment:
+      (REACT_NATIVE_DOGFOOD_APP_ENVIRONMENT as AppEnvironment) || 'demo',
     chatLabelNoted: false,
+    themeMode: 'dark',
+    useLocalSfu: false,
+    localIpAddress: '127.0.0.1',
   },
-  ['apiKey', 'appEnvironment', 'userId', 'userName', 'userImageUrl', 'appMode'],
+  [
+    'apiKey',
+    'appEnvironment',
+    'userId',
+    'userName',
+    'userImageUrl',
+    'appMode',
+    'themeMode',
+  ],
 );